2.2 Create synchronized call between services (creating)

+++ song-service/src/main/java/com/ms/songs/model/SongMetadata.java
@@ -21,7 +21,7 @@ public class SongMetadata {
 
     private String album;
 
-    private Integer length;
+    private Double length;
 
     private Integer year;
 }
